Convert scripts to Typescript
